requirements.txt
gunicorn==19.9.0
psycopg2-binary==2.7.6.1
django-request-logging==0.6.6